real_estate_dashboard/backend/requirements.txt
# Core Framework
fastapi==0.109.2
httpx==0.25.2
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0
requests==2.32.5

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Processing
numpy==1.26.2
numpy-financial==1.0.0
pandas==2.1.4

# YFinance & Economics Data
yfinance==0.2.66
multitasking==0.0.11
curl-cffi==0.13.0
frozendict==2.4.7
peewee==3.18.3
platformdirs==4.5.0
protobuf==6.33.0
websockets==15.0.1

# CLI & UI
rich==13.7.0

# Task Queue
celery==5.3.4

# LLM Integration
tenacity==8.2.3
redis==5.0.1

# Machine Learning & AI (Required by ML endpoints)
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
scipy==1.11.4

# Time Series & Forecasting (Required by analytics)
statsmodels==0.14.1

# Natural Language Processing (TEMPORARILY DISABLED - Heavy dependencies)
# transformers==4.45.0
# huggingface-hub==0.25.2
# torch==2.1.2
# sentence-transformers==3.3.0

# Document Generation & Export
reportlab==4.0.7
python-pptx==0.6.23
jinja2==3.1.2
pillow==10.1.0
matplotlib==3.8.2
weasyprint==60.1
